Normalizer Bug Fixes

+++ Stylometry/Normalizer.cs
@@ -19,10 +19,10 @@ namespace Stylometry
                     new TrainedData(
                     item.WordCount,
                     item.AverageLetterCount,
-                    item.NounFrequency / item.WordCountWithoutStopWords,
-                    item.VerbFrequency / item.WordCountWithoutStopWords,
-                    item.MostCommonWordCount / item.WordCount,
-                    item.TagsDiversity / item.WordCount,
+                    (float) item.NounFrequency / item.WordCountWithoutStopWords,
+                    (float) item.VerbFrequency / item.WordCountWithoutStopWords,
+                    (float) item.MostCommonWordCount / item.WordCount,
+                    (float) item.TagsDiversity / item.WordCount,
                     item.AuthorId)
                     );
             }
